pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["danaabbadi <abbadidana91@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
django = "^3.1.1"
djangorestframework = "^3.12.1"
djangorestframework_simplejwt = "^4.4.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
